+++ src/server/lagManueltBrevHtml.tsx
@@ -2,17 +2,8 @@ import * as React from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
 import { IManueltBrev } from '../typer/dokumentApi';
 import { Brevhode } from './components/Brevhode';
+import { dagensDatoFormatert } from '../utils/dato';
 import css from './utils/css';
-
-export const datoFormat = {
-  day: '2-digit',
-  month: '2-digit',
-  year: 'numeric',
-} as const;
-
-export const dagensDatoFormatert = (): string => {
-  return new Date().toLocaleDateString('no-NO', datoFormat);
-};
 
 export const lagManueltBrevHtml = (brev: IManueltBrev) => {
   return renderToStaticMarkup(
