+++ src/utils/dato.ts
@@ -1,0 +1,9 @@
+export const datoFormat = {
+  day: '2-digit',
+  month: '2-digit',
+  year: 'numeric',
+} as const;
+
+export const dagensDatoFormatert = (): string => {
+  return new Date().toLocaleDateString('no-NO', datoFormat);
+};
